pyproject.toml
[project]
name = "microecomerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-pymongo>=3.0.1",
    "python-dotenv>=1.2.1",
]
